Fix false negative caused by username changes
When someone changes their username, until Twitter updates the search
index, searches for their old username will return their tweets prior to
the change. This may cause a false negative for the search ban (and the
thread ban, which is not even checked in this case) if someone else
starts using the old username. I've added a check to make sure only
tweets actually sent by the same user are considered.

Closes #22.